Creación AvailabilityController
Modificación de AvailabilityService como interfaz y luego su implementación
